- exception voor negatieve inventory - rollack voor de exception - foutmelding op de htmlpagina voor de exception (WIP: laadt nieuwe pagina, zou gewoon moeten getoond worden maar nog niet gevonden)
