+++ eindwerkJava2/src/main/java/com/example/eindwerkJava2/controller/MutationController.java
@@ -1,5 +1,6 @@
 package com.example.eindwerkJava2.controller;
 
+import com.example.eindwerkJava2.Exceptions.NegativeInventoryException;
 import com.example.eindwerkJava2.model.Article;
 import com.example.eindwerkJava2.model.Category;
 import com.example.eindwerkJava2.model.Mutation;
@@ -44,20 +45,30 @@ public class MutationController {
 
     @GetMapping("/new/mutation")
     public String showNewMutationForm(Model model){
-     Mutation mutation = new Mutation();
+        Mutation mutation = new Mutation();
 
-     model.addAttribute("mutation",new Mutation());
-     model.addAttribute("transactiontypeList", transactionService.getTransactiontypes());
-     model.addAttribute("userList",userService.getActiveUsers());
-     model.addAttribute("articleList", articleService.getActiveArticles());
-     return "/forms/form_mutation";
+        model.addAttribute("mutation",new Mutation());
+        model.addAttribute("transactiontypeList", transactionService.getTransactiontypes());
+        model.addAttribute("userList",userService.getActiveUsers());
+        model.addAttribute("articleList", articleService.getActiveArticles());
+        return "/forms/form_mutation";
     }
 
     @PostMapping("/saveMutation")
-    public String saveMutation(@ModelAttribute("mutation") Mutation mutation){
-        this.mutationService.addMutation(mutation);
+    public String saveMutation(@ModelAttribute("mutation") Mutation mutation, Model model) throws NegativeInventoryException {
+        try {
+            this.mutationService.addMutation(mutation);
+        } catch (NegativeInventoryException ex){
+            model.addAttribute("mutation",new Mutation());
+            model.addAttribute("transactiontypeList", transactionService.getTransactiontypes());
+            model.addAttribute("userList",userService.getActiveUsers());
+            model.addAttribute("articleList", articleService.getActiveArticles());
+            model.addAttribute("errorMessage", ex.getMessage());
+            return "/forms/form_mutation";
+        }
         return "redirect:/mutation";
     }
+
 
     @GetMapping("delete/mutation/{mutationId}")
     public String deleteArticle(@PathVariable("mutationId") Long mutationId){
